+++ fixtures/rules/CG-CFG-021/bad/Parser.java
@@ -1,0 +1,6 @@
+import javax.xml.parsers.DocumentBuilderFactory;
+public class Parser {
+  public static Object parse() throws Exception {
+    return DocumentBuilderFactory.newInstance();
+  }
+}