+++ fixtures/rules/CG-CFG-021/good/Parser.java
@@ -1,0 +1,8 @@
+import javax.xml.parsers.DocumentBuilderFactory;
+public class Parser {
+  public static DocumentBuilderFactory newSafe() throws Exception {
+    DocumentBuilderFactory f = DocumentBuilderFactory.newInstance();
+    f.setFeature("http://apache.org/xml/features/disallow-doctype-decl", true);
+    return f;
+  }
+}